[refactor] devide code file by classes
