+++ TestShell_Excutor/TestShell_Excutor.vcxproj.filters
@@ -45,6 +45,12 @@
     <ClCompile Include="ShellTestScenarios.cpp">
       <Filter>소스코드</Filter>
     </ClCompile>
+    <ClCompile Include="ShellCommand.cpp">
+      <Filter>소스코드</Filter>
+    </ClCompile>
+    <ClCompile Include="ShellCommand_test.cpp">
+      <Filter>테스트파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -83,6 +89,12 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="ShellWrite.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="TS_function.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="ShellCommand.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
